Move deep_explore_schema from project to deep_explore
