+++ deep/schema.py
@@ -22,6 +22,8 @@ from notification import schema as notification_schema, mutation as notification
 from assisted_tagging import schema as assisted_tagging_schema
 from unified_connector import schema as unified_connector_schema
 from export import schema as export_schema, mutation as export_mutation
+from deep_explore import schema as deep_explore_schema
+
 from deep.enums import CustomEnum
 
 
@@ -36,6 +38,7 @@ class Query(
     lead_public_schema.Query,
     unified_connector_schema.Query,
     export_schema.Query,
+    deep_explore_schema.Query,
     # --
     graphene.ObjectType
 ):
